requirements.txt
absl-py==1.4.0
aiohttp==3.9.5
aiosignal==1.3.1
array_record==0.5.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
clu==0.0.12
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.8.0
etils==1.5.2
filelock==3.15.4
flatbuffers==24.3.25
flax==0.7.5
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
h5py==3.11.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
jax==0.4.23
jaxlib==0.4.23+cuda11.cudnn86
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.9.1
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.3.2
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.2.0.82
nvidia-cufft-cu11==10.9.0.58
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
opt-einsum==3.3.0
optax==0.2.2
optree==0.12.1
orbax-checkpoint==0.4.8
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
promise==2.3
protobuf==4.25.3
psutil==6.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-ternary==1.0.8
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==2.8.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-datasets==4.9.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.13.1
tensorstore==0.1.63
termcolor==2.4.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tqdm==4.66.4
transformers==4.42.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.4
Werkzeug==2.1.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
